split localization script into 2
