+++ .pipelines/pipeline.user.windows.buddy.yml
@@ -66,7 +66,11 @@ build:
         - from: 'out\loc'
           to: 'loc'
           include:
-            - '**/*' 
+            - '**/*'
+    - !!buildcommand
+      name: "Generate localized template package"
+      command: 'scripts\generateTemplates.cmd'
+      fail_on_stderr: true
     - !!defaultcommand
       name: 'Bump Version'
       command: 'scripts\bumpVersion.cmd'
