+++ .pipelines/pipeline.user.windows.yml
@@ -67,6 +67,10 @@ build:
           to: 'loc'
           include:
             - '**/*' 
+    - !!buildcommand
+      name: "Generate localized template package"
+      command: 'scripts\generateTemplates.cmd'
+      fail_on_stderr: true
     - !!defaultcommand
       name: 'Bump Version'
       command: 'scripts\bumpVersion.cmd'
