+++ .vscode/launch.json
@@ -22,6 +22,16 @@
     //     ],
     //     "type": "pwa-node"
     // }
+        // {
+        //     "type": "pwa-node",
+        //     "request": "launch",
+        //     "name": "Launch Program",
+        //     "skipFiles": [
+        //         "<node_internals>/**"
+        //     ],
+        //     "program": "${workspaceFolder}\\scripts\\extractStrings.js",
+        //     "args": ["C:\\src\\Application-Insights-Workbooks", "root"]
+        // },
         {
             "type": "pwa-node",
             "request": "launch",
@@ -29,8 +39,8 @@
             "skipFiles": [
                 "<node_internals>/**"
             ],
-            "program": "${workspaceFolder}\\scripts\\extractStrings.js",
-            "args": ["C:\\src\\Application-Insights-Workbooks", "root", "createPackage"]
+            "program": "${workspaceFolder}\\scripts\\generateTemplates.js",
+            "args": ["C:\\src\\Application-Insights-Workbooks", "root"]
         }
     ]
 }
