--- scripts/processTemplates.cmd
@@ -1,2 +1,0 @@
-powershell -NoProfile -ExecutionPolicy Unrestricted -Command "& '%~dp0processTemplates.ps1'"
-exit /B %ERRORLEVEL%
